[Refactor] Movido lógica de horário disponível do Service para o modelo Sala e removido o Service
